pyproject.toml
[tool.poetry]
name = "iiif-tiler-py"
version = "0.1.0"
description = ""
authors = ["Richard Benjamin Allen <richard@palaeopi.org.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.1"
imageio = "^2.34.0"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

